Change trust region behaviour of RandomLearner
... as a joke. Now it uses the trust region to determine how far it will
go from the last guess, rather than how far it will go from the best.
Note that the behaviour before this change was indeed how far it'll go
from the best, even though it doesn't look like it (due to it using the
value returned from put_params_and_get_cost) -- for RandomLearner the
elements of the cost queue are the current best parameters.

This change is essentially to document that behaviour, and play with
adding a test.