Change timing.py to be more consistent with nnlearner
It's still weird. timing.py suggests that running the epoch inside the
grpah is quicker, but with nnlearner I didn't see that. Maybe I didn't
look hard enough with nnlearner? Or maybe a bunch of little things just
added up? In any case, the difference doesn't seem to be particularly
big so probably isn't worth worrying about for now.