Switch bias std to 1
I *think* this will cause a bit more stochasticity in initial+trained
nets. Specifically, if we have a lower bias then things near the origin
(small x) will always have low activation, which means they can't train
fast. Similarly things away from the origin will have high activation
and hence train fast. This is causing the center to be "lazy" and the
edges to be "eager". A way this manifests itself is that if the cost
landscape has its minimum <0 and its max >0 (and the minimum is vaguely
central) then the edges will tend to shoot off to infinity (maybe legit,
although it will inhibit exploration) and the center will tend to hang
back towards 0 (bad, because for the center we definitely want it to err
on the side of low cost in order to encourage exploration).

If that's all right, using a higher bias will give more randomness to
the networks, so for both central and outer costs it will sometimes
overestimate and sometimes underestimate (and this is probably what we
want).