Use different BBB correlation method
We create a few buckets to store the epsilons. Each time we want a new
variable, we pick a random bucket. If there's an epsilon with the right
shape in that bucket we just use that, otherwise we create a new epsilon
to use (and put it in the bucket).