Add notebook for uncorrelated uncertainty
Turns out it *is* possibly to get the sorts of things we're looking for
(fitting the data well, and a large uncertainty away from the data). The
way it works is that the net creates a high-confidence sub-network that
fits the data well, and then has multiple high-confidence sub-networks
that are zero on the data but vary wildly away from it. The full net
uses the first sub-network with high confidence and the other nets with
low confidence.

Note that the other nets are individually high confidence (at least near
the data), because they need to be consistently zero on the data in
order to be included in the full net with low confidence.