More benchmarking results
Note that 10_3 and 6_3 were obtained with a slight learner modification.
Specifically, we don't train every net for each new data point, and we
don't cycle through nets in order. Instead for each generation we pick a
random net, train it if we haven't trained it since getting new data,
and sample from it.